refactor: move Constants.kt from amethyst to commons
Move relay Constants object to commons/src/commonMain/ with a
backward-compatible typealias in amethyst.
